backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
neo4j==5.24.0
google-genai>=1.0.0
httpx==0.27.2
python-dotenv==1.0.1
pandas==2.2.3
openpyxl==3.1.5
apscheduler==3.10.4
pydantic==2.9.2
pydantic-settings==2.5.2
pytest==8.3.3
pytest-asyncio==0.24.0
